refactor(signal,can): publish RAW u32 on signal bus; physical via CanDb_DecodeSignal
Change Signal_Get/Set/IsValid storage to u32; signal bus now carries

raw bit-patterns (NOT physical). Consumers (modules) derive physical

by calling CanDb_DecodeSignal(raw, sig_desc) on demand.

- app/signal: signal.{c,h} rewrite; u32 value, u32 Signal_Get(), file-

  level contract docs

- app/drv_api/can/can_db_codec.{h,c}: new CanDb_GetRaw() bridges

  is_signed to BitExtract / BitExtractSigned; bit-pattern safe round-trip

- app/drv_api/can/can_db.{c,h}: CanDb_DispatchByDb publishes raw

- can_if.c / can_rx.c: drop spurious (int32_t) casts on u32 -> Signal_Set

- mod_template.c / mod_can_demo.c: update local variables / printf formats

- docs/DBC_CHANGE_GUIDE.md, docs/SOC_SIGNAL_CONTRACT.md: u32 / raw text

+++ app/can/can_rx.c
@@ -378,9 +378,9 @@ static void prv_check_timeouts(void)
             map_hi2 |= ((u32)1u << (bit - 64u));
         }
     }
-    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_LO,  (int32_t)map_lo);
-    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_HI,  (int32_t)map_hi);
-    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_HI2, (int32_t)map_hi2);
+    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_LO,  map_lo);
+    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_HI,  map_hi);
+    (void)Signal_Set(SIG_CAN_RX_TIMEOUT_MAP_HI2, map_hi2);
 }
 
 /**
